Move properties over to `GameHost` for headless tests to work
The options themselves are still disabled on mobile platforms so this
should be no-op.
